Feature branch nr

+++ vite-project/src/journal-components/DeleteEntry.jsx
@@ -1,12 +1,17 @@
 import React from "react";
 import { useDeleteEntryMutation } from "../store/journalSlice";
+import "./journal-background.css";
 
 const DeleteEntry = ({ id, navigate }) => {
   const [deleteEntry, { isLoading, error }] = useDeleteEntryMutation();
 
   const handleDelete = async () => {
+    const confirmed = window.confirm(
+      "Are you sure you would like to delete this entry?"
+    );
+    if (!confirmed) return;
     try {
-        console.log("Delete button clicked")
+      console.log("Delete button clicked");
       await deleteEntry(id).unwrap();
       alert("Entry successfully deleted!");
       navigate("/journals");
@@ -17,11 +22,17 @@ const DeleteEntry = ({ id, navigate }) => {
   };
 
   return (
-    <div>
-      <button onClick={handleDelete} disabled={isLoading}>
-        {isLoading ? "Deleting Entry..." : "Delete"}
-      </button>
-      {error && <p>{error.message}</p>}
+    <div className="journal-background">
+      <div className="journal-content-wrapper">
+        <button
+          className="create-entry-button"
+          onClick={handleDelete}
+          disabled={isLoading}
+        >
+          {isLoading ? "Deleting Entry..." : "Delete"}
+        </button>
+        {error && <p>{error.message}</p>}
+      </div>
     </div>
   );
 };
